refactor: Remove `GroupByPartitioned` and dispatch to streaming engine
Fixes #24890.

This PR removes the `GroupByPartitionedExec` which used to be a fast path for
the in-memory. Now, if the same path path is detected, it will dispatch to the
streaming engine group-by.

This does not change any code to determine the fast paths, but this can be done
later if wanted.
