+++ crates/polars-mem-engine/src/planner/lp.rs
@@ -13,7 +13,8 @@ use self::python_dsl::PythonScanSource;
 use super::*;
 use crate::ScanPredicate;
 use crate::executors::{
-    self, CachePrefiller, Executor, PartitionedSinkExecutor, SinkExecutor, sink_name,
+    self, CachePrefiller, Executor, GroupByStreamingExec, PartitionedSinkExecutor, SinkExecutor,
+    sink_name,
 };
 use crate::predicate::PhysicalColumnPredicates;
 
@@ -246,6 +247,7 @@ fn create_physical_plan_impl(
             lp_arena.get(root),
             IR::Scan { .. } // Needed for the streaming impl
                 | IR::Cache { .. } // Needed for plans branching from the same cache node
+                | IR::GroupBy { .. } // Needed for the streaming impl
                 | IR::Sink { // Needed for the streaming impl
                     payload: SinkTypeIR::Partition(_),
                     ..
@@ -634,7 +636,7 @@ fn create_physical_plan_impl(
             keys,
             aggs,
             apply,
-            schema,
+            schema: _,
             maintain_order,
             options,
         } => {
@@ -694,27 +696,33 @@ fn create_physical_plan_impl(
                         false
                     }
                 });
+                let builder =
+                    build_streaming_executor.expect("invalid build. Missing feature new-streaming");
+
                 let input = recurse!(input, state)?;
-                let keys = keys
-                    .iter()
-                    .map(|e| e.to_expr(expr_arena))
-                    .collect::<Vec<_>>();
-                let aggs = aggs
-                    .iter()
-                    .map(|e| e.to_expr(expr_arena))
-                    .collect::<Vec<_>>();
-                Ok(Box::new(executors::PartitionGroupByExec::new(
+                let executor = Box::new(GroupByStreamingExec::new(
                     input,
+                    builder,
+                    root,
+                    lp_arena,
+                    expr_arena,
                     phys_keys,
                     phys_aggs,
                     maintain_order,
-                    options.slice,
-                    input_schema,
-                    schema,
+                    _slice,
                     from_partitioned_ds,
-                    keys,
-                    aggs,
-                )))
+                ));
+
+                // Use cache so that this runs during the cache pre-filling stage and not on the
+                // thread pool, it could deadlock since the streaming engine uses the thread
+                // pool internally.
+                let mut prefill = executors::CachePrefill::new_sink(executor);
+                let exec = prefill.make_exec();
+                let existing = cache_nodes.insert(prefill.id(), prefill);
+
+                assert!(existing.is_none());
+
+                Ok(Box::new(exec))
             } else {
                 let input = recurse!(input, state)?;
                 Ok(Box::new(executors::GroupByExec::new(
